+++ crates/parser/tests/macro.rs
@@ -273,7 +273,10 @@ fn macro_with_arg_calls() {
             Statement { ty: StatementType::Opcode(Opcode::Dup1), span: AstSpan(vec![Span { start: 351, end: 354, file: None }]) },
             Statement { ty: StatementType::Opcode(Opcode::Dup3), span: AstSpan(vec![Span { start: 418, end: 421, file: None }]) },
             Statement { ty: StatementType::Opcode(Opcode::Gt), span: AstSpan(vec![Span { start: 492, end: 493, file: None }]) },
-            Statement { ty: StatementType::ArgCall("error".to_string()), span: AstSpan(vec![Span { start: 566, end: 570, file: None }]) },
+            Statement {
+                ty: StatementType::ArgCall("TRANSFER_TAKE_FROM".to_string(), "error".to_string()),
+                span: AstSpan(vec![Span { start: 566, end: 570, file: None }]),
+            },
             Statement { ty: StatementType::Opcode(Opcode::Jumpi), span: AstSpan(vec![Span { start: 573, end: 577, file: None }]) },
             Statement { ty: StatementType::Opcode(Opcode::Dup2), span: AstSpan(vec![Span { start: 715, end: 718, file: None }]) },
             Statement { ty: StatementType::Opcode(Opcode::Swap1), span: AstSpan(vec![Span { start: 780, end: 784, file: None }]) },
@@ -564,7 +567,11 @@ fn macro_invocation_with_arg_call() {
             Statement {
                 ty: StatementType::MacroInvocation(MacroInvocation {
                     macro_name: "TRANSFER_TAKE_FROM".to_string(),
-                    args: vec![MacroArg::ArgCall("error".to_string(), AstSpan(vec![Span { start: 92, end: 92, file: None }]))],
+                    args: vec![MacroArg::ArgCall(ArgCall {
+                        macro_name: "ARG_CALL".to_string(),
+                        name: "error".to_string(),
+                        span: AstSpan(vec![Span { start: 92, end: 92, file: None }]),
+                    })],
                     span: AstSpan(vec![
                         Span { start: 67, end: 84, file: None },
                         Span { start: 85, end: 85, file: None },
@@ -586,7 +593,11 @@ fn macro_invocation_with_arg_call() {
             Statement {
                 ty: StatementType::MacroInvocation(MacroInvocation {
                     macro_name: "TRANSFER_GIVE_TO".to_string(),
-                    args: vec![MacroArg::ArgCall("error".to_string(), AstSpan(vec![Span { start: 126, end: 126, file: None }]))],
+                    args: vec![MacroArg::ArgCall(ArgCall {
+                        macro_name: "ARG_CALL".to_string(),
+                        name: "error".to_string(),
+                        span: AstSpan(vec![Span { start: 126, end: 126, file: None }]),
+                    })],
                     span: AstSpan(vec![
                         Span { start: 103, end: 118, file: None },
                         Span { start: 119, end: 119, file: None },