+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Update dependencies to the latest version.
 - Fix nested macro argument resolution (fixes #60, #48).
 - Fix label resolution for macros passed as arguments (fixes #62).
+- Fix macro argument scoping and execution order (fixes #50).
 
 ## [1.1.12] - 2025-05-30
 - Clean release after CI fixes.